added check for values from access_token.content, removed duplication in gitignore and renamed to formAPI in event controller

+++ packages/admin-api/src/controllers/notification.controller.ts
@@ -1,21 +1,25 @@
+/* eslint-disable camelcase */
 import express from "express"
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import * as notificationService from "../services/notification.service"
 import { getCatchments } from "../lib/catchment"
 
-export const getNotification = async (req: any, res: express.Response) => {
+export const getNotifications = async (req: any, res: express.Response) => {
     try {
         const { catchmentNo, type } = req.query
         const catchments = await getCatchments(req.kauth.grant.access_token)
         if (catchments.length === 0 || !catchments.includes(Number(catchmentNo))) {
             return res.status(403).send("Forbidden")
         }
-        const { email } = req.kauth.grant.access_token.content
         if (type !== "application" && type !== "claim") {
-            return res.status(400).send("Invalid type")
+            return res.status(400).send("Bad Request")
         }
-        const username = `${req.kauth.grant.access_token.content.idp_username}@${req.kauth.grant.access_token.content.identity_provider}`
-        const notifications = await notificationService.getNotification(email, Number(catchmentNo), type, username)
+        const { email, idp_username, identity_provider } = req.kauth.grant.access_token.content
+        if (email === undefined || idp_username === undefined || identity_provider === undefined) {
+            return res.status(403).send("Forbidden")
+        }
+        const username = `${idp_username}@${identity_provider}`
+        const notifications = await notificationService.getNotifications(email, Number(catchmentNo), type, username)
         return res.status(200).send(notifications)
     } catch (e: unknown) {
         console.error(e)
@@ -32,11 +36,14 @@ export const addNotification = async (req: any, res: express.Response) => {
             return res.status(403).send("Forbidden")
         }
 
-        const { email } = req.kauth.grant.access_token.content
         if (type !== "application" && type !== "claim") {
-            return res.status(400).send("Invalid type")
+            return res.status(400).send("Bad Request")
         }
-        const username = `${req.kauth.grant.access_token.content.idp_username}@${req.kauth.grant.access_token.content.identity_provider}`
+        const { email, idp_username, identity_provider } = req.kauth.grant.access_token.content
+        if (email === undefined || idp_username === undefined || identity_provider === undefined) {
+            return res.status(403).send("Forbidden")
+        }
+        const username = `${idp_username}@${identity_provider}`
         const result = await notificationService.addNotification(email, Number(catchmentNo), type, username)
         return res.status(200).send(result)
     } catch (e: unknown) {
@@ -53,12 +60,14 @@ export const deleteNotification = async (req: any, res: express.Response) => {
         if (catchments.length === 0 || !catchments.includes(Number(catchmentNo))) {
             return res.status(403).send("Forbidden")
         }
-
-        const { email } = req.kauth.grant.access_token.content
         if (type !== "application" && type !== "claim") {
-            return res.status(400).send("Invalid type")
+            return res.status(400).send("Bad Request")
         }
-        const username = `${req.kauth.grant.access_token.content.idp_username}@${req.kauth.grant.access_token.content.identity_provider}`
+        const { email, idp_username, identity_provider } = req.kauth.grant.access_token.content
+        if (email === undefined || idp_username === undefined || identity_provider === undefined) {
+            return res.status(403).send("Forbidden")
+        }
+        const username = `${idp_username}@${identity_provider}`
         const result = await notificationService.deleteNotification(email, Number(catchmentNo), type, username)
         return res.status(200).send(result)
     } catch (e: unknown) {
@@ -69,6 +78,10 @@ export const deleteNotification = async (req: any, res: express.Response) => {
 
 export const checkNotificationEmail = async (req: any, res: express.Response) => {
     try {
+        const { email, idp_username, identity_provider } = req.kauth.grant.access_token.content
+        if (email === undefined || idp_username === undefined || identity_provider === undefined) {
+            return res.status(403).send("Forbidden")
+        }
         const username = `${req.kauth.grant.access_token.content.idp_username}@${req.kauth.grant.access_token.content.identity_provider}`
         const notifications = await notificationService.getAllNotificationsFromUser(username)
         await Promise.all(
